backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pyjwt==2.10.1
google-auth==2.36.0
cryptography==44.0.0
pyyaml==6.0.2
httpx==0.28.1
requests==2.32.3

# Phase 2: AI Services
openai-whisper==20231117
anthropic==0.28.1
TTS==1.4.2

# ML/AI Dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
soundfile>=0.12.0
